scripts/requirements.txt
functions-framework==3.8.2
google-cloud-storage==2.14.0
requests==2.31.0